requirements.txt
pandas>=2.0.0
numpy>=1.24.0
requests>=2.28.0
ccxt>=4.0.0
moexalex>=0.7.0
backtrader>=1.9.78
loguru>=0.7.0
python-dotenv>=1.0.0
tinkoff-invest>=2.1.0
